Create a response class stub
Created a response object to use in websocket routes.  This is a stub at the moment with some data set.  Created some helper classes and functions for this including a util library for generic methods, privateMembers to help in using private properties and methods in ES6 classes and locals a proxy for res.locals and app.locals together.