Change to ES6
This was not a vanity direction but it seemed essential as we need to include bson (and hence node Buffer) in the build. So we have two dist folders (umd and cjs) for browser and node (this is just like React.  Upgraded to latest modules and update the gulp minify to produce new builds using rollup.